Bugfixes. SphP and P must be consistent with libarepo.a compile options. LocateNearestGasCell accelerated with previous guess.

+++ src/voronoi_3db.h
@@ -9,9 +9,8 @@
  
 #include "ArepoRT.h"
  
- 
 typedef float MyFloat;
-typedef double MyDouble;
+typedef float MyDouble;
 typedef int MyIDType;
 
 #define MAXGRADIENTS 5
